+++ src/models/dprnn.py
@@ -2,7 +2,7 @@ import torch
 import torch.nn as nn
 
 from utils.model import choose_rnn
-from utils.utils_tasnet import choose_layer_norm
+from utils.tasnet import choose_layer_norm
 
 EPS=1e-12
